MiniPlayer Revision for Radiant Player 2.0
Finally the MiniPlayer revision is here!! (Due note this is still in
its early stages; this has been tested heavily.)

Features:
- Complete re-design of the MiniPlayer
- RadiantTools (an extension tool for GMusic.js)
- MiniPlayer Visualizations
- Volume support
- Drop menu for options
- Artwork animations
- Dock / Undock the MiniPlayer
- IMFL (I'm Feeling Lucky ~ Roll the Dice)
- Youtube w/ Fullscreen support
- Drag support in an undocked state
- Replay back 10 seconds / Forward 30 seconds actions

Remember this is in its early stages, if you encounter any issues
please report them! I did my best to kick out any issues making this as
smooth as possible.

This is going to be continuously worked on with other planned features
in the future.